Switch the public_id and internal_id around
The public_id is now stored as a "public_id" file, where the
"internal_id" is stored as the directory path

This makes renaming scripts far easier as the directory does
not need moving.

It also removes the O(N) operations when trying to resolve a
script from the `internal_id`.

There is still a O(N) operation when trying to resolve a script
from the `public_id`, however this can be negated by referencing
script's by the `internal-id`