Unbreak viewing unsubmited jobs
Previously, `job view-job-stdout` for a job that failed to submit
resulted in `bin/job: no implicit conversion of nil into String`.

This resulted in the API `FlightJobScriptAPI::CommandError -
Unexpectedly failed to load file: SsDGhXIH.stdout` and responding with a
`503` FSR.

The client responded to the `503` with a generic "something went wrong"
error message and looked completely broken.

This commit fixes `job view-job-stdout` to print a help message and
exit `20`.