Change how session script is created
Previously, the job script itself loaded the flight environment (i.e.,
`flight start`) and any modules required for the workload (e.g., `flight
env activate gridware; module load apps/R`).  Those would set a number
of environment variables which may or may not be available to the
desktop session.

Flight Desktop is user-configurable as to whether environment variables
are propagated to the session.  If the user has elected not export
environment variables to the session, the launched app will not work.

This commit fixes that issue by creating a session script which loads
the flight environment, and any modules.  As they are done as part of
the session script, we can be certain that they will be available to
launch the application.

The code has been organized so that the plumbing functions can later be
hidden away once we have more advanced editing of scripts.