Job ends when session ends
There is no need for the job to continue after the desktop session has
ended.  We should terminate the job and free up the resources as soon as
possible.

The `wait_session` function has been extended to sleep until the job's
desktop session is no longer "Active".

This requires the job to be able to obtain it's desktop session id.  If
it knows its own flight job id,  it can do this via `flight job info-job
--wait-desktop $JOB_ID`.  This is now exposed to the job script via
`$FLIGHT_JOB_ID`.