Add tags for templates and scripts
Instead of having a growing number adhoc attributes added to templates,
scripts and jobs, we now have a "tags" attribute.

A use of tags that has not yet been implemented would be to support
filtering across tags.  Something like: `flight job list-scripts
--filter script:type=interactive` to list only interactive scripts.

The format for tags is a list of tripple tags. See
https://en.wikipedia.org/wiki/Tag_(metadata)#Triple_tags

When a script is created, its tags are copied from the template.
Currently, tags are not copied from the script to the job.