Remove `includes` from config file
I can't see any use case for a valid global include value.  Suggesting
to users that they might want to configure a valid global setting is
likely to do nothing but confuse.

Instead let's have this as an "undocumented feature" that we might
remove at some point in the future.