WIP: thoughts on validating answers
This commit contains thoughts on how validations could be added to
questions.  The validations are going to be ran against the answer
provided by the user.  The ideas here are more concerned with what is
presented to the template author than how the validation would be
implemented.

The validations are added in a new `validate` key, before I explain
that, let me clarify the purpose of the existing `format` key.

The `format` key is intended to specify the way in which the question is
formatted and presented to the user.  Currently, 4 variations are
supported:

* `text`: a single line of text analogous to HTML's `<input
  type="text">`.
* `multiline_text`: multiple lines of text analogous to HTML's
  `<textarea>`.
* `select`: a single answer from a predefined selection analogous to
  HTML's `<select>`.
* `multiselect`: multiple answers from a predefined selection analogous
  to HTML's `<select multiple>`.

Other variations could be added that are analogous to HTML's input
types, `number`, `email`, `date`, `time`, etc..

Whilst some of the HTML inputs provide some validation on the provided
value, the purpose of the questions `format` key is to format the
question for the user, not to validate the answer.

To validate the answer, a new `validate` key is to be added.  It's
format is influenced by https://json-schema.org/.  We are already using
JSON Schema to validate the templates themselves, so using it to
validate the answer should be relatively simple.

A limited set of functionality will initially be supported.  Examples of
the validations that I'm initially thinking of are below:

Validate the "type" of the answer.  This will check that the provided
answer can be parsed to the expected type.

```
validate:
  type: string|number|integer
```

Validate that an answer has been provided:

```
validate:
  required: true
```

Validate against a set of acceptable values:

```
validate:
  enum:
    - '14.04.013'
    - '13.06.012'
```

Validate against a regular expression:

```
validate:
  type: string
  pattern: "\\d{2}:\\d{2}"
```

Validate arrays of items:

```
validate:
  type: array
  items:
    enum: ['BEGIN', 'END', 'FAIL', 'ALL']
```

Validate ranges of numbers, or just a minimum or maximum bound:

```
validate:
  type: integer
  minimum: 1
  maximum: 100
```