Ensure `QuestionPrompter` never ignores answers
Previously, if a user provided answers with the `--answers` flag and
FSR, the `QuestionPrompter` need to be ran, the answers would be
discarded.  The `QuestionPrompter` would show a summary of the default
answers, and the default answers would be used in the creation of the
script.

Now, the answers provided with `--answers` are not discarded.  The
`QuestionPrompter` uses them in the summary and they are used in the
creation of the script.