Cancelled array jobs don't break listing jobs
Previously, if a task metadata file was saved as
`jobs/simple-array-5/tasks/metadata.[1-10%4].yaml` the `list-jobs`
command was broken. As was the `info-job` command was broken for that
array job.

The creation of that file may only have been possible with a "between
commits" code-state, but we now defensively code around it.