Improve robustness with submitting jobs
The error paths in which a job can be submitted to the HPC scheduler and
Flight Job be unable to track that job have been greatly reduced.

After the job has been submitted to the HPC scheduler and we have a
scheduler ID, the job record is persisted before any other information
about that job is collected.