Fix listing scripts when template is unknown
Previously, there was an issue where listing scripts could fail if the
template was not known.

The exact command that caused this was `flight_JOB_includes=template
flight job list-scripts --json`.  Which is the command that the API
runs.

This is fixed by ensuring that `Template#metadata` always returns a hash
and never `false`.