Permit jobs submitted via flight-desktop/srun
Lots of work to decouple job metadata from our original mechanism of
submitting jobs via `sbatch`.

Previously, we required that all metadata for a job was present in the
`metadata.yaml` file when it was loaded.  Now we allow metadata to be
present in some control files and merge it into the job's metadata prior
to validating the metadata.

This new mechanism means that desktop/srun submitted jobs can create
control files to record the job's `scheduler_id`, `job_type` and
`submit_status`.