Avoid unsetting scheduler id for interactive jobs
If an interactive job starts immediately, the scheduler id control is
set when the job script runs on the client.  It is then unset when the
`srun` script completes without printing out the scheduler's id.

If the job has not been processed by Flight Job in the mean time this
can result in Flight Job "losing"  the job's scheduler id.

The fix to this is to not write empty values to control files.  There
are no valid situations where an emtpy value is valid or where we'd like
to unset a control file.