Remove duplicated `let` in specs
The duplicated `let` blocks can be moved to a common scope.  That they
depend on a `let(:job_id) { ... }` that differs for each spec is not an
issue.

When the `let` block is evaluated it will be done in the context of a
particular spec and each spec has a context defining a `let(:job_id)
{...}`.