Change specs testing metadata content
Previously, the specs were testing that a existing script has the
attributes present in its metadata file.  This is a useful test, but
already exists in the `reads #{attr} from the metadata` specs.  If the
wrong value was written to the metadata file, FSR, these specs would not
detect that.

Now, the spec checks that the existing script has the attribute that we
*expect* to be in the metadata file.  This will catch errors in what is
written to the metadata file.