Fix bug with interactive jobs on login nodes
Previously, there was a race condition for interactive jobs on login
nodes.  The control file `job_type` would be updated to `BOOTSTRAPPING`
then `srun` ran.  When either `srun` reported that the job was pending
or the job started running, the control file `scheduler_id` and would be
updated.  The job was in an invalid state between the `job_type` and
`scheduler_id` file being written.

Now, `job_type` and `scheduler_id` control files are written when either
`srun` reports that the job was pending or the job starts running.

The race condition is still present, as the writting of the two files is
not atomic.  Future work can consider how this could be addressed.