Kill desktop sessions when job terminates
Desktop sessions running inside of a scheduler allocation were not being
terminated when the scheduler allocation was deallocated (aka the
scheduler job terminated).  This was observed for job cancellation and
may have also been the behaviour for other ways in which the job could
terminate.

This commit fixes the issue by adding a `session_term` function and
calling it when the job script receives `SIGTERM`.  The `session_term`
function kills the desktop session if it exists.